+++ src/index.js
@@ -1,10 +1,11 @@
 const { loadEnv, loadConfig } = require("./config");
 const { syncAll } = require("./sync");
+const { fetchAll } = require("./fetch");
 const { uploadAsset, updateAsset } = require("./upload");
 const { generateLuau, generateDts } = require("./codegen");
-const { EXT_TO_ASSET_TYPE } = require("./sync");
+const { EXT_TO_ASSET_TYPE, CONVERTED_EXT_TO_ASSET_TYPE } = require("./sync");
 const { watchAll } = require("./watch");
-const { fetchDecalImageId, extractImageIdFromAssetBody } = require("./image-id");
+const { fetchAssetContent, fetchDecalImageId, extractImageIdFromAssetBody } = require("./image-id");
 const { buildAssetMap, loadLockForSync, lockPathForSync, normalizeAssetPath, resolveEntryAssetId } = require("./asset-map");
 const { DEFAULT_CODEGEN_FORMAT, listCodegenFormats, registerCodegenFormat, resolveCodegenFormat } = require("./formats");
 const {
@@ -26,13 +27,16 @@ module.exports = {
 	loadConfig,
 	syncAll,
 	watchAll,
+	fetchAll,
 	uploadAsset,
 	updateAsset,
+	fetchAssetContent,
 	fetchDecalImageId,
 	extractImageIdFromAssetBody,
 	generateLuau,
 	generateDts,
 	EXT_TO_ASSET_TYPE,
+	CONVERTED_EXT_TO_ASSET_TYPE,
 	DEFAULT_CODEGEN_FORMAT,
 	buildAssetMap,
 	buildStudioPluginManifest,